+++ docs/conf.py
@@ -188,6 +188,7 @@ nitpick_ignore = [
     ("py:class", "substra.Client"),
     ("py:class", "substra.sdk.client.Client"),
     ("py:class", "substra.sdk.models.ComputePlan"),
+    ("py:class", "substra.sdk.schemas.AlgoCategory"),
     ("py:class", "ComputePlan"),
     ("py:class", "substratools.algo.CompositeAlgo"),
     ("py:class", "substratools.algo.AggregateAlgo"),