+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- BREAKING CHANGE: convert (test task) to (predict task + test task) (#217)
+
 ## [0.20.0](https://github.com/owkin/connectlib/releases/tag/0.20.0) - 2022-07-05
 
 - compatibility with substra 0.28.0